kvm_vm.py: Handling generated spice_port
If wrong spice_port in test A is set and in test B which folows test A is expected auto-generated spice_port, Autotest uses wrong value from previous test. That's because all other parameters are the same. For this reason is added value 'generate' to mark spice_port for generate new value.
This new value is now handled in kvm_vm.py in qemu_kvm command and in migration.

Corrected some pylint warnings

Signed-off-by: Vaclav Ehrlich <vehrlich@redhat.com>